requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=1.0.0=pypi_0
astunparse=1.6.3=pypi_0
biopython=1.81=pypi_0
blas=1.0=mkl
bottleneck=1.3.4=py38hce1f21e_0
ca-certificates=2022.9.14=ha878542_0
cachetools=5.0.0=pypi_0
certifi=2022.9.14=pyhd8ed1ab_0
charset-normalizer=2.0.12=pypi_0
compress-pickle=2.1.0=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
dm-haiku=0.0.6=pypi_0
dm-sonnet=2.0.0=pypi_0
dm-tree=0.1.6=pypi_0
et-xmlfile=1.1.0=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.39.4=pypi_0
gast=0.5.3=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jax=0.3.4=pypi_0
jaxlib=0.3.2+cuda11.cudnn805=pypi_0
jmp=0.0.2=pypi_0
joblib=1.1.0=pyhd8ed1ab_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=13.0.0=pypi_0
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
liblapack=3.9.0=12_linux64_mkl
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.6=pypi_0
matplotlib=3.7.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ml-collections=0.1.1=pypi_0
ncurses=6.3=h7f8727e_2
numexpr=2.8.1=py38h6abb31d_0
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.2.0=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1q=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38h295c915_0
pillow=9.5.0=pypi_0
pip=21.2.4=py38h06a4308_0
protobuf=3.19.4=pypi_0
pyaml=21.10.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2022.1=py38h06a4308_0
pyyaml=6.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scikit-optimize=0.9.0=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.12.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.11=h1ccaba5_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.9=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.0=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
